+++ packages/util/src/string.ts
@@ -1,15 +1,15 @@
-const isBlank = (s: string) => {
-	return s.split('').every((char) => {
+const isBlank = (string: string) => {
+	return string.split('').every((char) => {
 		return char === ' ';
 	});
 };
 
-const isEmpty = (s: string) => {
-	return !s;
+const isEmpty = (string: string) => {
+	return !string;
 };
 
-const capitalize = (s: string) => {
-	return s.charAt(0).toUpperCase() + s.slice(1);
+const capitalize = (string: string) => {
+	return string.charAt(0).toUpperCase() + string.slice(1);
 };
 
 export { capitalize, isBlank, isEmpty };