+++ packages/util/CHANGELOG.md
@@ -1,3 +1,9 @@
-## 0.0.0 (1 Oct 2023)
+## 0.0.1 (2 Oct 2023)
+
+-   (Fix) Use free function for `isFalse`, `isFalsy` and `isTruthy`
+-   (Fix) Pass boolean value to Bool property function
+-   (Fix) Allow functions to be imported
+
+## 0.0.0 (2 Oct 2023)
 
 -   (Feat) Initial public release